+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQXQTYP50W5KN2H3MZDE6SD6"
+        "01JQYPMWQHD57SBKSH17RQ7ZMQ"
     )
     r"""The id of the resource"""