+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQXQTYP3Z6FV35HMJKZ1SNDY"
+        "01JQYPMWQFCYY75G59GZ5CRZNE"
     )
     r"""The id of the resource"""